backend/requirements.txt
absl-py==2.3.1
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
executing==2.1.0
Faker==32.1.0
fastjsonschema==2.20.0
Flask==3.0.3
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.55.3
fqdn==1.5.1
gast==0.6.0
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==3.1.1
grpcio==1.74.0
h11==0.9.0
h2==3.2.0
h5py==3.14.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
namex==0.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.2
opencv-contrib-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
PyAudio==0.2.14
pycparser==2.22
Pygments==2.18.0
pymongo==4.10.1
pyodbc==5.2.0
pyparsing==3.2.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyttsx3==2.99
pytz==2024.2
pywin32==307
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==14.1.0
rpds-py==0.20.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.6
SpeechRecognition==3.14.3
SQLAlchemy==2.0.36
srt==3.5.3
stack-data==0.6.3
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
textblob==0.19.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
vosk==0.3.45
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
zipp==3.21.0
